chore: revert changes outside of noxfile
