+++ tests/system_tests/test_artifacts/test_registry.py
