+++ wandb/apis/public/artifacts.py
@@ -59,6 +59,7 @@ from wandb.sdk.artifacts._generated import (
     RunOutputArtifactsProjectRunOutputArtifacts,
 )
 from wandb.sdk.artifacts._graphql_fragments import omit_artifact_fields
+from wandb.sdk.artifacts._validators import validate_artifact_name
 from wandb.sdk.internal.internal_api import Api as InternalApi
 from wandb.sdk.lib import deprecate
 
@@ -290,7 +291,7 @@ class ArtifactCollection:
         self.client = client
         self.entity = entity
         self.project = project
-        self._name = name
+        self._name = validate_artifact_name(name)
         self._saved_name = name
         self._type = type
         self._saved_type = type
@@ -430,8 +431,8 @@ class ArtifactCollection:
         return self._name
 
     @name.setter
-    def name(self, name: List[str]) -> None:
-        self._name = name
+    def name(self, name: str) -> None:
+        self._name = validate_artifact_name(name)
 
     @property
     def type(self):
